# internal/api/definition/access_token_body.go
// Code generated by go-swagger; DO NOT EDIT.

package definition

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// AccessTokenBody access token body
//
// swagger:model AccessTokenBody
type AccessTokenBody struct {

	// access token
	AccessToken string `json:"access_token,omitempty"`
}

// Validate validates this access token body
func (m *AccessTokenBody) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this access token body based on context it is used
func (m *AccessTokenBody) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *AccessTokenBody) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *AccessTokenBody) UnmarshalBinary(b []byte) error {
	var res AccessTokenBody
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}